RestaurantsDetail이 포함된 파일명 일부를 RestaurantDetail으로 변경
하나의 레스토랑에 대한 상세페이지이기 떄문에 `Restaurants`에서 `Restaurant`으로
변경해주었습니다
