+++ src/RestaurantDetail.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-export default function RestaurantsDetail({ restaurant }) {
+export default function RestaurantDetail({ restaurant }) {
   return (
     <>
       {restaurant ? (
