+++ src/RestaurantDetailPage.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import RestaurantDetailContainer from './RestaurantDetailContainer';
+
+function RestaurantDetailePage() {
+  return (
+    <>
+      <h2>레스토랑 리스트</h2>
+      <RestaurantDetailContainer />
+    </>
+  );
+}
+
+export default RestaurantDetailePage;
