--- src/RestaurantsDetailPage.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import RestaurantsDetailContainer from './RestaurantsDetailContainer';
-
-function RestaurantsDetailePage() {
-  return (
-    <>
-      <RestaurantsDetailContainer />
-    </>
-  );
-}
-
-export default RestaurantsDetailePage;
